requirements.txt
absl-py==0.9.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
blis==0.4.1
Bottleneck==1.3.2
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
decorator==4.4.2
deepcut==0.7.0.0
defusedxml==0.6.0
dill==0.3.2
emoji==0.5.4
entrypoints==0.3
fastai==1.0.61
fastprogress==0.2.3
fasttext==0.9.2
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
hyperopt==0.2.4
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lightgbm==2.3.1
marisa-trie==0.7.5
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
notebook==6.0.3
numexpr==2.7.1
numpy==1.19.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pythainlp==2.1.4
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
sklearn-crfsuite==0.3.6
soupsieve==2.0.1
spacy==2.3.1
srsly==1.0.2
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tinydb==4.1.1
tokenizers==0.8.1rc1
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
transformers==3.0.2
urllib3==1.25.9
wasabi==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
